web_ui/requirements.txt
# Web UI requirements for MCP Document Exchange System
flask>=2.3.0
werkzeug>=2.3.0
python-dotenv>=1.0.0

# Additional dependencies for the MCP clients
mcp>=1.0.0
asyncio-mqtt>=0.14.0
